pyproject.toml
[project]
name = "taskflow-backend"
version = "0.1.0"
description = "TaskFlow — Modular Event-Driven Task Management Backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "celery>=5.4",
  "redis>=5.0",
  "email-validator>=2.0",
  "bcrypt==4.0.1",
  "passlib==1.7.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
